added interfaces to class and created types

+++ src/modules/pass-collection/pass-collection.db.service.ts
@@ -1,6 +1,5 @@
-import { Injectable, BadRequestException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
-import { handlers } from 'src/handlers/handlers';
 import {
   ICreateReq,
   ICreateRes,
@@ -12,8 +11,8 @@ import {
   IGetByIdReq,
   IGetByIdRes,
 } from './pass-collection.types';
+import { handlerErrorDb } from 'src/handlers/handlerError.db';
 
-// SERVICE
 interface IPassCollectionDbService {
   findAll(data: IGetAllReq): Promise<IGetAllRes[]>;
   findById(data: IGetByIdReq): Promise<IGetByIdRes>;
@@ -28,73 +27,35 @@ export class PassCollectionDbService implements IPassCollectionDbService {
   constructor(private readonly databaseService: DatabaseService) {}
 
   public async findAll({ userId }: IGetAllReq): Promise<IGetAllRes[]> {
-    if (!userId) throw new BadRequestException();
-
-    const res = await handlers.dbError(
-      this.databaseService.passCollection.findMany({
-        where: { userId },
-      }),
-    );
-
-    return res.map((item) => {
-      if (userId === item.userId) return { id: item.id, name: item.name };
-    });
+    const res = await handlerErrorDb(this.databaseService.passCollection.findMany({ where: { userId } }));
+    return res.map((item) => ({ id: item.id, name: item.name }));
   }
 
   public async findById({ id, userId }: IGetByIdReq): Promise<IGetByIdRes> {
-    if (!userId || !id) throw new BadRequestException();
-
-    const res = await handlers.dbError(
-      this.databaseService.passCollection.findFirst({
-        where: { userId, id },
-      }),
-    );
-
-    if (res.userId === userId) return { id: res.id, name: res.name, data: res.data };
-    return null;
+    const res = await handlerErrorDb(this.databaseService.passCollection.findFirst({ where: { userId, id } }));
+    return { id: res.id, name: res.name, data: res.data };
   }
 
   public async create({ userId, name, data }: ICreateReq): Promise<ICreateRes> {
-    if (!userId || !name || !data) throw new BadRequestException();
-
-    const res = await handlers.dbError(
-      this.databaseService.passCollection.create({
-        data: { userId, name, data },
-      }),
-    );
-
+    const res = await handlerErrorDb(this.databaseService.passCollection.create({ data: { userId, name, data } }));
     return { id: res.id };
   }
 
   public async editName({ id, userId, name }: IEditNameReq): Promise<void> {
-    if (!id || !userId || !name) throw new BadRequestException();
-
-    await handlers.dbError(
-      this.databaseService.passCollection.update({
-        where: { id, userId },
-        data: { name },
-      }),
-    );
+    await handlerErrorDb(this.databaseService.passCollection.update({ where: { id, userId }, data: { name } }));
   }
 
   public async editData({ id, userId, data }: IEditDataReq): Promise<void> {
-    if (!id || !userId || !data) throw new BadRequestException();
-
-    await handlers.dbError(
-      this.databaseService.passCollection.update({
-        where: { id, userId },
-        data: { data },
-      }),
-    );
+    await handlerErrorDb(this.databaseService.passCollection.update({ where: { id, userId }, data: { data } }));
   }
 
   public async delete({ id, userId }: IDeleteInDbReq): Promise<void> {
-    if (!id || !userId) throw new BadRequestException();
-
-    await handlers.dbError(
-      this.databaseService.passCollection.delete({
-        where: { id, userId },
-      }),
-    );
+    await handlerErrorDb(this.databaseService.passCollection.delete({ where: { id, userId } }));
   }
+
+  // private checkRequiredField(value: unknown, fieldName: string): void {
+  //   if (!value) {
+  //     throw new BadRequestException(`${fieldName} is required.`);
+  //   }
+  // }
 }
